requirements.txt
selenium==4.16.0
opencv-python==4.8.1
pillow==10.1.0
tensorflow==2.15.0
numpy==1.24.3
webdriver-manager==4.0.1
python-dotenv==1.0.0
scikit-learn==1.3.0
joblib==1.3.2
modelcontextprotocol-sdk==0.1.0
aiohttp==3.9.1
